aventureiro

+++ CartasSuperTrunfo.c
@@ -46,7 +46,7 @@ int main() {
 
     // Cálculo da densidade populacional e PIB per capita da carta 1
     densidade_populacional1 = (float)populacao1 / area1;
-    pib_per_capita1 = pib1 / populacao1;
+    pib_per_capita1 = (float)pib1 / populacao1;
 
     // Entrada de dados da carta 2
     printf("\nCarta 2\n");
@@ -74,7 +74,7 @@ int main() {
 
     // Cálculo da densidade populacional e PIB per capita da carta 2
     densidade_populacional2 = (float)populacao2 / area2;
-    pib_per_capita2 = pib2 / populacao2;
+    pib_per_capita2 = (float)pib2 / populacao2;
 
     // Exibir os dados da carta 1
     printf("\n-----Carta 1-----\n");
